+++ variables.tf
@@ -45,3 +45,121 @@ variable "extra_tags" {
   type        = map(string)
   default     = {}
 }
+
+variable "deny_all_inbound" {
+  description = "True to deny all inbound traffic by default"
+  type        = bool
+  default     = true
+}
+
+variable "http_inbound_allowed" {
+  description = "True to allow inbound HTTP traffic"
+  type        = bool
+  default     = false
+}
+
+variable "allowed_http_source" {
+  description = "Allowed source for inbound HTTP traffic. Can be a Service Tag, \"*\" or a CIDR list."
+  type        = any
+  default     = []
+  validation {
+    condition = (
+      var.allowed_http_source != null &&
+      var.allowed_http_source != "" &&
+      (can(tolist(var.allowed_http_source)) || can(tostring(var.allowed_http_source)))
+    )
+    error_message = "Variable must be a Service Tag, \"*\" or a list of CIDR."
+  }
+}
+
+variable "https_inbound_allowed" {
+  description = "True to allow inbound HTTPS traffic"
+  type        = bool
+  default     = false
+}
+
+variable "allowed_https_source" {
+  description = "Allowed source for inbound HTTPS traffic. Can be a Service Tag, \"*\" or a CIDR list."
+  type        = any
+  default     = []
+  validation {
+    condition = (
+      var.allowed_https_source != null &&
+      var.allowed_https_source != "" &&
+      (can(tolist(var.allowed_https_source)) || can(tostring(var.allowed_https_source)))
+    )
+    error_message = "Variable must be a Service Tag, \"*\" or a list of CIDR."
+  }
+}
+
+variable "ssh_inbound_allowed" {
+  description = "True to allow inbound SSH traffic"
+  type        = bool
+  default     = false
+}
+
+variable "allowed_ssh_source" {
+  description = "Allowed source for inbound SSH traffic. Can be a Service Tag, \"*\" or a CIDR list."
+  type        = any
+  default     = []
+  validation {
+    condition = (
+      var.allowed_ssh_source != null &&
+      var.allowed_ssh_source != "" &&
+      (can(tolist(var.allowed_ssh_source)) || can(tostring(var.allowed_ssh_source)))
+    )
+    error_message = "Variable must be a Service Tag, \"*\" or a list of CIDR."
+  }
+}
+
+variable "rdp_inbound_allowed" {
+  description = "True to allow inbound RDP traffic"
+  type        = bool
+  default     = false
+}
+
+variable "allowed_rdp_source" {
+  description = "Allowed source for inbound RDP traffic. Can be a Service Tag, \"*\" or a CIDR list."
+  type        = any
+  default     = []
+  validation {
+    condition = (
+      var.allowed_rdp_source != null &&
+      var.allowed_rdp_source != "" &&
+      (can(tolist(var.allowed_rdp_source)) || can(tostring(var.allowed_rdp_source)))
+    )
+    error_message = "Variable must be a Service Tag, \"*\" or a list of CIDR."
+  }
+}
+
+variable "winrm_inbound_allowed" {
+  description = "True to allow inbound WinRM traffic"
+  type        = bool
+  default     = false
+}
+
+variable "allowed_winrm_source" {
+  description = "Allowed source for inbound WinRM traffic. Can be a Service Tag, \"*\" or a CIDR list."
+  type        = any
+  default     = []
+  validation {
+    condition = (
+      var.allowed_winrm_source != null &&
+      var.allowed_winrm_source != "" &&
+      (can(tolist(var.allowed_winrm_source)) || can(tostring(var.allowed_winrm_source)))
+    )
+    error_message = "Variable must be a Service Tag, \"*\" or a list of CIDR."
+  }
+}
+
+variable "application_gateway_rules_enabled" {
+  description = "True to configure rules mandatory for hosting an Application Gateway. See https://docs.microsoft.com/en-us/azure/application-gateway/configuration-infrastructure#allow-access-to-a-few-source-ips"
+  type        = bool
+  default     = false
+}
+
+variable "load_balancer_rules_enabled" {
+  description = "True to configure rules mandatory for hosting a Load Balancer."
+  type        = bool
+  default     = false
+}